+++ src/main/java/com/back/domain/news/real/controller/RealNewsController.java
@@ -1,10 +1,9 @@
-package com.back.domain.news.realNews.controller;
+package com.back.domain.news.real.controller;
 
-import com.back.domain.news.realNews.dto.NaverNewsDto;
-import com.back.domain.news.realNews.dto.RealNewsDto;
-import com.back.domain.news.realNews.service.RealNewsService;
-import com.back.domain.news.util.NewsPageService;
-import com.back.domain.news.util.NewsType;
+import com.back.domain.news.real.dto.RealNewsDto;
+import com.back.domain.news.real.service.RealNewsService;
+import com.back.domain.news.common.service.NewsPageService;
+import com.back.domain.news.common.enums.NewsType;
 import com.back.global.rsData.RsData;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
