+++ src/main/java/com/back/domain/news/real/service/RealNewsService.java
@@ -1,10 +1,10 @@
-package com.back.domain.news.realNews.service;
+package com.back.domain.news.real.service;
 
-import com.back.domain.news.realNews.dto.NaverNewsDto;
-import com.back.domain.news.realNews.dto.NewsDetailDto;
-import com.back.domain.news.realNews.dto.RealNewsDto;
-import com.back.domain.news.realNews.entity.RealNews;
-import com.back.domain.news.realNews.repository.RealNewsRepository;
+import com.back.domain.news.common.dto.NaverNewsDto;
+import com.back.domain.news.common.dto.NewsDetailDto;
+import com.back.domain.news.real.dto.RealNewsDto;
+import com.back.domain.news.real.entity.RealNews;
+import com.back.domain.news.real.repository.RealNewsRepository;
 import com.back.global.util.HtmlEntityDecoder;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
