+++ src/main/java/com/back/domain/news/today/entity/TodaysNews.java
@@ -1,0 +1,4 @@
+package com.back.domain.news.today.entity;
+
+public class TodaysNews {
+}
